requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
python-docx==0.8.11
pandas==2.1.1
python-multipart==0.0.6
